% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edxr.R
\name{ChildNames}
\alias{ChildNames}
\title{get child names of an xml node}
\usage{
ChildNames(node)
}
\arguments{
\item{node}{xml tree node}
}
\value{
character vector of child names
}
\description{
get child names of an xml node
}
\details{
get child names of an xml node
}
\examples{
test_01_path <- system.file("extdata", "week_01_test_01_ibcorr.xml", package = "edxr")
doc <- XML::xmlTreeParse(test_01_path)
v_num <- GetVersion(doc, 10, 1)
ChildNames(v_num)
}

