% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edxr.R
\name{GetNumberOfQuestions}
\alias{GetNumberOfQuestions}
\title{obtain number of a questions from coursera xml}
\usage{
GetNumberOfQuestions(doc)
}
\arguments{
\item{doc}{parsed xml tree, obtained by XML::xmlTreeParse}
}
\value{
character test title
}
\description{
obtain number of a questions from coursera xml
}
\details{
obtain number of a questions from coursera xml
}
\examples{
test_01_path <- system.file("extdata", "week_01_test_01_ibcorr.xml", package = "edxr")
doc <- XML::xmlTreeParse(test_01_path)
GetNumberOfQuestions(doc)
}

