% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edxr.R
\name{GetOption}
\alias{GetOption}
\title{get option from version}
\usage{
GetOption(version, option_group_no, option_no)
}
\arguments{
\item{version}{xml tree of a version}

\item{option_group_no}{int number of option group}

\item{option_no}{int number of option}
}
\value{
xml tree with version of option group
}
\description{
get option from version
}
\details{
get option from version
}
\examples{
test_01_path <- system.file("extdata", "week_01_test_01_ibcorr.xml", package = "edxr")
doc <- XML::xmlTreeParse(test_01_path)
q_06_1 <- GetVersion(doc, 6, 1)
option <- GetOption(q_06_1, 1, 1)
}

