% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edxr.R
\name{GetOptionGroup}
\alias{GetOptionGroup}
\title{get option group from version}
\usage{
GetOptionGroup(version, option_group_no)
}
\arguments{
\item{version}{xml tree of a version}

\item{option_group_no}{int number of option group}
}
\value{
xml tree with version of option group
}
\description{
get option group from version
}
\details{
get option group from version
}
\examples{
test_01_path <- system.file("extdata", "week_01_test_01_ibcorr.xml", package = "edxr")
doc <- XML::xmlTreeParse(test_01_path)
q_06_1 <- GetVersion(doc, 6, 1)
og <- GetOptionGroup(q_06_1, 1)
}

