% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edxr.R
\name{GetPreamble}
\alias{GetPreamble}
\title{Obtain preamble of a test from coursera xml}
\usage{
GetPreamble(doc)
}
\arguments{
\item{doc}{parsed xml tree, obtained by XML::xmlTreeParse}
}
\value{
character test title
}
\description{
Obtain preamble of a test from coursera xml
}
\details{
Obtain preamble of a test from coursera xml
}
\examples{
test_01_path <- system.file("extdata", "week_01_test_01_ibcorr.xml", package = "edxr")
doc <- XML::xmlTreeParse(test_01_path)
GetPreamble(doc)
}

