% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edxr.R
\name{GetQuestion}
\alias{GetQuestion}
\title{obtain all versions of a question from coursera xml}
\usage{
GetQuestion(doc, question_no)
}
\arguments{
\item{doc}{parsed xml tree, obtained by XML::xmlTreeParse}

\item{question_no}{int number of question}
}
\value{
xml tree with question
}
\description{
obtain all versions of a question from coursera xml
}
\details{
obtain all versions of a question from coursera xml
}
\examples{
test_01_path <- system.file("extdata", "week_01_test_01_ibcorr.xml", package = "edxr")
doc <- XML::xmlTreeParse(test_01_path)
q_17 <- GetQuestion(doc, 17)
}

