% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edxr.R
\name{GetType}
\alias{GetType}
\title{get type of a version (numeric/checkbox/radio)}
\usage{
GetType(version)
}
\arguments{
\item{version}{xml tree of version}
}
\value{
character type of version
}
\description{
get type of a version (numeric/checkbox/radio)
}
\details{
get type of a version (numeric/checkbox/radio)
}
\examples{
test_01_path <- system.file("extdata", "week_01_test_01_ibcorr.xml", package = "edxr")
doc <- XML::xmlTreeParse(test_01_path)
v_num <- GetVersion(doc, 10, 1)
GetType(v_num)
}

