% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/edxr.R
\name{GetVersion}
\alias{GetVersion}
\title{obtain specific version of a question from coursera xml}
\usage{
GetVersion(doc, question_no, version_no)
}
\arguments{
\item{doc}{parsed xml tree, obtained by XML::xmlTreeParse}

\item{question_no}{int number of question}

\item{version_no}{int number of version}
}
\value{
xml tree with version of question
}
\description{
obtain specific version of a question from coursera xml
}
\details{
obtain specific version of a question from coursera xml
}
\examples{
test_01_path <- system.file("extdata", "week_01_test_01_ibcorr.xml", package = "edxr")
doc <- XML::xmlTreeParse(test_01_path)
v_num <- GetVersion(doc, 10, 1)
v_checkbox <- GetVersion(doc, 4, 1)
v_radio <- GetVersion(doc, 3, 1)
}

